backend/requirements.txt
django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9python-telegram-bot==20.8
